Add placeholder for dribbble posts
We should have a placeholder that fades out when the dribbble posts are
loading, especially important for first loads.

There are a few different approaches to what that placeholder should be.
I feel like have one solid web page layout placeholder works because
we're loading what will likely be product designs. The alternative would
be something like 6 empty boxes that are the same size as the incoming
dribbble posts. I like this layout thing though.